Refactored implementation of account management, roles and staff/student creation

+++ usermanagement/src/main/java/com/example/usermanagement/repository/RoleRepository.java
@@ -9,15 +9,18 @@ import org.springframework.stereotype.Repository;
 import com.example.usermanagement.model.Role;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Long> {
+public interface RoleRepository extends JpaRepository<Role, Integer> {
     List<Role> findByRoleNameIn(List<String> roleNames);
-    Optional <Role> findByRoleName(String roleName);
-    Optional <Role> findByDescription(String description);
-    Optional <Role> findByRoleId(Integer userId);
+
+    Optional<Role> findByRoleName(String roleName);
+
+    Optional<Role> findByDescription(String description);
+
+    Optional<Role> findByRoleId(Integer userId);
 
     void deleteByRoleId(Integer userId);
 
     boolean existsByRoleName(String roleName);
+
     boolean existsByDescription(String Description);
 }
-
